Move playlist members feature to its own package
