+++ app/src/main/java/fr/nihilus/music/library/playlists/members/MembersHolder.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Thibault Seisel
+ * Copyright 2022 Thibault Seisel
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package fr.nihilus.music.library.playlists
+package fr.nihilus.music.library.playlists.members
 
 import android.graphics.Bitmap
 import android.support.v4.media.MediaBrowserCompat.MediaItem
